move 'note' to 'book controller.
