+++ 17_Spring_Jpa_Jquery_JS/src/main/java/ru/otus/spring/jquery/controller/BookController.java
@@ -5,11 +5,18 @@ import org.springframework.web.bind.annotation.*;
 import ru.otus.spring.jquery.domain.Author;
 import ru.otus.spring.jquery.domain.Book;
 import ru.otus.spring.jquery.domain.Genre;
+import ru.otus.spring.jquery.domain.Note;
 import ru.otus.spring.jquery.dto.BookDto;
+import ru.otus.spring.jquery.dto.NoteDto;
+import ru.otus.spring.jquery.dto.mapper.NoteMapper;
 import ru.otus.spring.jquery.services.AuthorService;
 import ru.otus.spring.jquery.services.BookService;
 import ru.otus.spring.jquery.services.GenreService;
+import ru.otus.spring.jquery.services.NoteService;
+
 import java.util.List;
+
+import static java.util.stream.Collectors.toList;
 import static org.springframework.http.HttpStatus.CREATED;
 
 
@@ -21,6 +28,8 @@ public class BookController {
     private final BookService service;
     private final AuthorService authorService;
     private final GenreService genreService;
+    private final NoteService noteService;
+    private final NoteMapper mapper;
 
     @GetMapping("/books")
     public List<Book> finAll() {
@@ -56,6 +65,15 @@ public class BookController {
     @DeleteMapping("/books/{id}")
     public void delete(@PathVariable("id") Long id) {
         service.deleteById(id);
+    }
+
+    @GetMapping("/book/{bookId}/notes")
+    public List<NoteDto> findByBookId(@PathVariable("bookId") Long bookId) {
+        List<Note> notes = noteService.findByBookId(bookId);
+        for (Note note : notes) {
+            System.out.println(note.getId() + " + " + note.getNote());
+        }
+        return noteService.findByBookId(bookId).stream().map(mapper::toDto).collect(toList());
     }
 
 }
