+++ 17_Spring_Jpa_Jquery_JS/src/main/resources/static/index.html
@@ -68,6 +68,7 @@
     </table>
 
     <br><br><br>
+    <!--TODO: would be better to make it accessible only dooring book editing..-->
     <span id="noteList">
         <table class="books" id="notes">
             <thead>
